+++ StockAPI/DTOs/UserTradeDTO.cs
@@ -2,7 +2,7 @@ namespace StockAPI.DTOs
 {
     public class UserTradeDTO
     {
-        public int TradeId { get; set; }
+        public long TradeId { get; set; }
         public string Username { get; set; }
         public string Symbol { get; set; }
         public decimal? Units { get; set; }
